pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-csv-crew"
version = "0.1.0"
description = "Hybrid Search RAG for CSV Data with Multi-Agent Orchestration"
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "NearForm", email = "info@nearform.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",  # File uploads

    # Database
    "psycopg[pool]>=3.1.0",  # PostgreSQL driver with connection pooling
    "pgvector>=0.2.4",  # Vector similarity search

    # Data Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # AI/ML
    "crewai>=0.11.0",  # Multi-agent orchestration
    "anthropic>=0.8.0",  # Claude Opus API
    "openai>=1.6.0",  # OpenAI embeddings
    "google-genai>=1.0.0",  # Google Gemini embeddings (alternative, replaces google-generativeai)
    "langchain-anthropic>=0.1.0",  # LangChain Anthropic integration
    "langchain-groq>=0.1.0",  # LangChain GROQ integration

    # Authentication
    "python-jose[cryptography]>=3.3.0",  # JWT tokens
    "passlib[bcrypt]>=1.7.4",  # Password hashing (for future use)

    # Utilities
    "python-dotenv>=1.0.0",  # Environment variables
    "chardet>=5.2.0",  # Character encoding detection
    "pandas>=2.1.0",  # CSV processing
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",  # Parallel testing
    "testcontainers>=3.7.0",  # PostgreSQL integration tests
    "httpx>=0.26.0",  # FastAPI test client

    # Quality Tools
    "ruff==0.8.4",  # Linting and formatting (pinned to match pre-commit hook version)
    "mypy>=1.8.0",  # Type checking
    "pylint>=3.0.3",  # Additional linting

    # Type Stubs
    "types-python-jose>=3.3.0",
    "types-passlib>=1.7.0",
    "types-requests>=2.31.0",
]

[project.scripts]
rag-csv-crew = "backend.src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "tests*"]
namespaces = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--cov=backend/src",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (database, external services)",
    "contract: API contract tests (OpenAPI validation)",
    "slow: Slow-running tests",
    "e2e: End-to-end tests with real API calls (disabled by default to avoid quota usage)",
]

[tool.coverage.run]
source = ["backend/src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["backend/src", "tests"]
extend-exclude = [
    ".venv",
    "venv",
    "__pycache__",
    "*.egg-info",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RET", # flake8-return
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate (commented-out code)
    "PL",  # pylint
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # FastAPI Depends() in defaults (standard FastAPI pattern)
    "B904",   # raise from pattern (will address case-by-case)
    "ARG002", # Unused method argument (parameters for future use)
    "PLR0913", # Too many arguments (we'll handle this case-by-case)
    "PLR2004", # Magic value used in comparison (acceptable for tests)
    "PLR0912", # Too many branches (existing code)
    "PLR0915", # Too many statements (existing code)
    "PLW0602", # Global variable warnings (existing code)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # Use of assert (required for pytest)
    "ARG001", # Unused function argument (fixtures)
    "PLR2004", # Magic value in comparison (acceptable in tests)
    "B017",   # pytest.raises(Exception) is acceptable in tests
    "SIM105", # contextlib.suppress can be less clear in tests
    "SIM117", # Nested with statements are fine in tests
]
"backend/src/api/*.py" = [
    "ARG001", # Unused function argument (FastAPI dependencies)
]

[tool.ruff.lint.isort]
known-first-party = ["backend", "tests"]
force-sort-within-sections = true

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "crewai.*",
    "anthropic.*",
    "google.genai",
    "langchain_anthropic.*",
    "langchain_groq.*",
    "testcontainers.*",
    "pgvector.*",
    "chardet",
]
ignore_missing_imports = true

[tool.pylint.main]
py-version = "3.13"
jobs = 0  # Use all available CPUs
load-plugins = [
    "pylint.extensions.bad_builtin",
    "pylint.extensions.code_style",
    "pylint.extensions.comparison_placement",
    "pylint.extensions.consider_ternary_expression",
    "pylint.extensions.docparams",
    "pylint.extensions.docstyle",
    "pylint.extensions.empty_comment",
    "pylint.extensions.eq_without_hash",
    "pylint.extensions.for_any_all",
    "pylint.extensions.mccabe",
    "pylint.extensions.overlapping_exceptions",
    "pylint.extensions.private_import",
    "pylint.extensions.redefined_loop_name",
    "pylint.extensions.set_membership",
    "pylint.extensions.typing",
]

[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "missing-module-docstring",  # We'll add these during implementation
    "R6103",  # consider-using-assignment-expr (walrus operator is optional style)
]

[tool.pylint.design]
max-args = 7
max-attributes = 10
max-bool-expr = 5
max-branches = 12
max-locals = 15
max-public-methods = 20
max-returns = 6
max-statements = 50
min-public-methods = 1

[tool.pylint.format]
max-line-length = 100

[tool.pylint.refactoring]
max-nested-blocks = 5

[tool.pylint.similarities]
min-similarity-lines = 4
ignore-comments = true
ignore-docstrings = true
ignore-imports = true

[dependency-groups]
dev = [
    "types-python-dateutil>=2.9.0.20260124",
]
